refactor(store):moving store definition from App to store index
