+++ src/store/index.ts
@@ -1,4 +1,7 @@
-import { combineReducers } from 'redux';
+import { combineReducers, createStore, applyMiddleware, compose } from 'redux';
+import createSagaMiddleware from 'redux-saga';
+import { createBrowserHistory } from 'history';
+import { rootSaga } from './sagas';
 import { userReducer } from './user/reducers';
 import { groupReducer } from './groups/reducers';
 import { predictionsReducer } from './predictions/reducers';
@@ -9,4 +12,14 @@ export const rootReducer = combineReducers({
 	prediction: predictionsReducer
 });
 
+export const browserHistory = createBrowserHistory();
+const sagaMiddleware = createSagaMiddleware();
+const composeEnhancers = (typeof window !== 'undefined' && (window as any).__REDUX_DEVTOOLS_EXTENSION_COMPOSE__) || compose;
+export const store = createStore(
+  rootReducer,
+  composeEnhancers(applyMiddleware(sagaMiddleware))
+)
+sagaMiddleware.run(rootSaga, browserHistory)
+
 export type RootState = ReturnType<typeof rootReducer>;
+
